backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
celery
redis
pillow
pypdf
docx2pdf
pandas
openpyxl
tesseract-ocr
